+++ stormpath/resources/account.py
@@ -1,6 +1,9 @@
 """Stormpath Account resource mappings."""
 
 
+from six import string_types
+
+from stormpath.error import Error as StormpathError
 from .base import (
     AutoSaveMixin,
     CollectionResource,
@@ -48,17 +51,28 @@ class Account(Resource, AutoSaveMixin, DeleteMixin, StatusMixin):
             'tenant': Tenant,
         }
 
-    def add_group(self, group_object_or_href_or_name):
-        """Associate a Group with this Account.
+    def _resolve_group(self, group_object_or_href_or_name):
+        """Given a Group object or href or name, return a functional Group
+        object.
 
-        This creates a
-        :class:`stormpath.resources.group_membership.GroupMembership` resource
-        on the backend.
+        This helper method allows us to easily accept Group arguments in
+        multiple ways.
 
-        :param group_object_or_href_or_name: This could be any one of the following:
-            - A :class:`stormpath.resources.group.Group` object.
-            - A Group href, ex: 'https://api.stormpath.com/v1/groups/3wzkqr03K8WxRp8NQuYSs3'
+        :param group_object_or_href_or_name: This could be any one of the
+            following:
+
+            - A :class:`stormpath.resources.group.Group` object, that already
+              exists in Stormpath.
+            - A Group href, ex:
+              'https://api.stormpath.com/v1/groups/3wzkqr03K8WxRp8NQuYSs3'
             - A Group name, ex: 'admins'.
+
+        :raises:
+            - ValueError if an invalid href or name is specified.
+            - TypeError if a non-Group object is specified.
+
+        :rtype: obj
+        :returns: A matching Group object.
 
         .. note::
             Passing in a :class:`stormpath.resources.group.Group` object will
@@ -67,53 +81,119 @@ class Account(Resource, AutoSaveMixin, DeleteMixin, StatusMixin):
         """
         from .group import Group
 
-        # First, we'll check to see whether or not this is a string.
-        group = group_object_or_href_or_name
-        if isinstance(group, basestring):
+        # If this is a Group object already, we have no work to do!
+        if isinstance(group_object_or_href_or_name, Group):
+            return group_object_or_href_or_name
+
+        # We now know this isn't a Group object.
+        href_or_name = group_object_or_href_or_name
+
+        # Check to see whether or not this is a string.
+        if isinstance(group_object_or_href_or_name, string_types):
 
             # If this Group is an href, we'll just use that.
-            if group.startswith('https://api.stormpath.com/'):
-                group = self.groups.get(group)
+            if href_or_name.startswith(self._client.BASE_URL):
+                href = href_or_name
+
+                try:
+                    group = self.directory.groups.get(href)
+
+                    # We're accessing group.name here to force evaluation of
+                    # this Group -- this allows us to check and see whether or
+                    # not this Group is actually valid.
+                    group.name
+                except StormpathError:
+                    raise ValueError('Invalid Group href specified.')
+
+                return group
 
             # Otherwise, we'll assume this is a Group name, and try to query
             # it.
             else:
-                for g in self.directory.groups.query(name=group):
-                    if g.name == group:
-                        group = g
-                        break
+                name = href_or_name
+                groups = self.directory.groups.query(name=name)
 
-        # If this is not a Group instance, something horrible was given to us,
+                for g in groups:
+                    if g.name == name:
+                        return g
+
+                raise ValueError('Invalid Group name specified.')
+
+        # If this is not a string instance, something horrible was given to us,
         # so bail.
-        elif not isinstance(group, Group):
-            raise TypeError('Unsupported type. Group object required.')
+        raise TypeError('Unsupported type. Group object, href, or name required.')
 
+    def add_group(self, group_object_or_href_or_name):
+        """Associate a Group with this Account.
+
+        This creates a
+        :class:`stormpath.resources.group_membership.GroupMembership` resource
+        on the backend.
+
+        :param group_object_or_href_or_name: This could be any one of the
+            following:
+
+            - A :class:`stormpath.resources.group.Group` object.
+            - A Group href, ex:
+                'https://api.stormpath.com/v1/groups/3wzkqr03K8WxRp8NQuYSs3'
+            - A Group name, ex: 'admins'.
+
+        .. note::
+            Passing in a :class:`stormpath.resources.group.Group` object will
+            always be the quickest way to add a Group, as it doesn't require
+            any additional API calls.
+        """
+        group = self._resolve_group(group_object_or_href_or_name)
         return self._client.group_memberships.create({
             'account': self,
             'group': group,
         })
 
-    def in_group(self, group):
+    def has_group(self, group_object_or_href_or_name):
         """Check to see whether or not this Account is a member of the
         specified Group.
 
-        :param group: A :class:`stormpath.resources.group.Group` object.
+        :param group_object_or_href_or_name: This could be any one of the
+            following:
+
+            - A :class:`stormpath.resources.group.Group` object.
+            - A Group href, ex:
+                'https://api.stormpath.com/v1/groups/3wzkqr03K8WxRp8NQuYSs3'
+            - A Group name, ex: 'admins'.
+
+        .. note::
+            Passing in a :class:`stormpath.resources.group.Group` object will
+            always be the quickest way to check a Group's membership, since it
+            doesn't require any additional API calls.
 
         :returns: True if the Account is a member of the Group, False
             otherwise.
         """
-        for g in self.groups.search(group.name):
+        group = self._resolve_group(group_object_or_href_or_name)
+
+        for g in self.groups.query(name=group.name):
             if g.name == group.name:
                 return True
 
         return False
 
-    def in_groups(self, groups, all=True):
+    def has_groups(self, group_objects_or_hrefs_or_names, all=True):
         """Check to see whether or not this Account is a member of a list of
         Groups.
 
-        :param groups: A list of :class:`stormpath.resources.group.Group`
-            objects to check.
+        :param group_objects_or_hrefs_or_names: A list of either:
+            - :class:`stormpath.resources.group.Group` objects.
+            - Group hrefs, ex:
+                'https://api.stormpath.com/v1/groups/3wzkqr03K8WxRp8NQuYSs3'
+            - Group names, ex: 'admins'.
+
+            This could look something like:
+            [
+                group,
+                'https://api.stormpath.com/v1/groups/3wzkqr03K8WxRp8NQuYSs3',
+                'admins',
+            ]
+
         :param all: A boolean (default: True) which controls how Group
             assertions are handled.  If all is set to True (default), then
             we'll check to ensure that this Account is a member of ALL Groups
@@ -124,8 +204,9 @@ class Account(Resource, AutoSaveMixin, DeleteMixin, StatusMixin):
         """
         total_checks = 0
 
+        groups = group_objects_or_hrefs_or_names
         for group in groups:
-            if self.in_group(group):
+            if self.has_group(group):
                 total_checks += 1
 
                 if not all:
@@ -143,6 +224,36 @@ class Account(Resource, AutoSaveMixin, DeleteMixin, StatusMixin):
         http://docs.stormpath.com/console/product-guide/#cloud-directory-workflow-automations
         """
         return self.get_status() == self.STATUS_UNVERIFIED
+
+    def remove_group(self, group_object_or_href_or_name):
+        """Remove this Account from the specified Group.
+
+        :param group_object_or_href_or_name: This could be any one of the
+            following:
+
+            - A :class:`stormpath.resources.group.Group` object.
+            - A Group href, ex:
+                'https://api.stormpath.com/v1/groups/3wzkqr03K8WxRp8NQuYSs3'
+            - A Group name, ex: 'admins'.
+
+        :raises: :class:`stormpath.error.StormpathError` if the Group specified
+            does not contain this Account.
+
+        .. note::
+            Passing in a :class:`stormpath.resources.group.Group` object will
+            always be the quickest way to check a Group's membership, since it
+            doesn't require any additional API calls.
+        """
+        group = self._resolve_group(group_object_or_href_or_name)
+
+        for membership in self.group_memberships:
+            if membership.group.href == group.href:
+                membership.delete()
+                return
+
+        raise StormpathError({
+            'developerMessage': 'This user is not part of Group %s.' % group.name,
+        })
 
 
 class AccountList(CollectionResource):