+++ CHANGES.md
@@ -1,7 +1,7 @@
 Version 1.1.0
 -------------
 
-Released on March 10, 2014.
+Not yet released.
 
 - Adding docs.
 - Adding customizable authentication methods to the Client.
@@ -10,6 +10,26 @@ Released on March 10, 2014.
   simpler.
 - Adding `Account.in_groups()` helper method, to make asserting multiple group
   memberships simpler.
+- Adding the ability for `Account.add_group()` to support Group objects, hrefs,
+  or names.
+- Adding the ability for `Account.has_group()` to support Group objects, hrefs,
+  or names.
+- Adding the ability for `Account.has_groups()` to support Group objects, hrefs,
+  or names.
+- Adding the ability for `Group.add_account()` to support Account objects, hrefs,
+  usernames or emails.
+- Adding new method, `Account.remove_group(group_object_or_href_or_name)`. This
+  lets users easily remove Groups from an Account.
+- Adding new method,
+  `Group.remove_account(account_object_or_href_or_username_or_email)`. This lets
+  users easily remove Accounts from a Group.
+- Adding new method,
+  `Group.has_account(account_object_or_href_or_username_or_email)`. This lets
+  users easily check to see whether an Account is in a Group.
+- Adding new method,
+  `Group.has_accounts(account_objects_or_hrefs_or_usernames_or_emails)`. This lets
+  users easily check to see whether a list of Accounts are part of a Group or
+  not.
 
 
 Version 1.0.1